Added Requirements class
Requirements allow the creator of a behavior to set the required parts
for robots that want to use the behavior. Not every part used in the
behavior needs to be set in the Requirements, this allows users to set
optional parts. When the yuse a part they should make sure to check if the part is null if it's not set in Requirements.