Changed min android version to 10
this change is due to the Visualizer Component used in the DanceBehavior.
If the device is older than version 9, it can't be used. If someone wants
to use RoboBrain for a lower android version, the code needs to handle
version stuff and DanceBehavior can not react to audio fft data